% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermutationGenerationMethods.R
\name{pseudorandom_generation}
\alias{pseudorandom_generation}
\title{Fast pseudorandom unique permutation generation}
\usage{
pseudorandom_generation()
}
\arguments{
\item{original}{an integer (0/1 only) or logical vector representing the
statuses of the true observations.}

\item{number_permutations}{integer; desired number of permutations to find.}
}
\value{
A 0/1 matrix with \code{number_permutations} rows and
\code{length(original)} columns. If there are fewer possible permutations
than are requested, an error may be thrown.
}
\description{
Called when \code{uniqueperm::get_unique_perms(method = "pseudo")}. It is the
fastest method in the package, but due to the way that permutations are
selected is not strictly random. If the permutations generated don't need to
be distinct from another set generated at a different time, use this option.
}
\details{
This method is a helper function not meant for general use. Use
\code{\link{get_unique_perms}} instead, as it has more parameter checking and
safeguards. For example, if there are fewer possible permutations than are
requested, calling this method directly will cause an error.
}
\examples{
example_states <- c(0,0,0,0,1,1,1,1)
pseudorandom_generation(example_states, 34)
get_unique_perms(example_states, 34, method = "pseudo")

example_states <- c(rep(1, 500), rep(0,500))
pseudorandom_generation(example_states, 1000)

}
\seealso{
Other permutation generation methods: \code{\link{bitset_generation}},
  \code{\link{fast_generation}}
}
\author{
Sage Wright, Samantha Jensen
}
